requirements.txt
Django==1.5.1
MySQL-python==1.2.3
South==0.7.6
URLObject==2.1.0
argparse==1.2.1
django-filter==0.6
djangorestframework==2.3.3
factory-boy==2.0.2
ipython==0.13.1
pep8==1.4.5
pyflakes==0.7
python-memcached==1.51
wsgiref==0.1.2
